+++ ai_ref_kits/multimodal_ai_visual_generator/config/illustration.yaml
