+++ ai_ref_kits/multimodal_ai_visual_generator/convert_and_optimize_text2image.py
