--- ai_ref_kits/multimodal_ai_visual_generator/index.html
@@ -1,53 +1,0 @@
-<html>
-  <head>
-    <meta charset="UTF-8">
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/pixi.js/5.1.3/pixi.min.js"></script>
-    <style>
-      body { background-color: black;
-      }
-      .center {
-        border: 1px solid;
-        margin: auto;
-        left: 5%;
-        top: 4%;
-        padding: 1px;
-        position: absolute;
-      }
-    </style>
-  </head>
-  <body>
-    <div id="test_app" style="width: 500px; height: 500px" class="center"></div>
-    
-    <script>
-      const IMG_WIDTH = 2304;
-      const IMG_HEIGHT = 1296;
-      const SCALE_FACTOR = 25;
-      
-      const app = new PIXI.Application({
-        width: IMG_WIDTH, 
-        height: IMG_HEIGHT
-      });
-      const container = document.getElementById("test_app");
-
-      container.appendChild(app.view);
-
-      const img = new PIXI.Sprite.from("results/original_image.png");
-      img.width = IMG_WIDTH;
-      img.height = IMG_HEIGHT;
-      app.stage.addChild(img);
-
-      depthMap = new PIXI.Sprite.from("results/depth_map.png");
-      depthMap.width = IMG_WIDTH;
-      depthMap.height = IMG_HEIGHT;
-      app.stage.addChild(depthMap);
-
-      displacementFilter = new PIXI.filters.DisplacementFilter(depthMap);
-      app.stage.filters = [displacementFilter];
-
-      container.addEventListener("mousemove", (e) => {
-        displacementFilter.scale.x = (IMG_WIDTH / 2 - e.clientX) / SCALE_FACTOR;
-        displacementFilter.scale.y = (IMG_HEIGHT / 2 - e.clientY) / SCALE_FACTOR;
-      })
-    </script>
-  </body>
-</html>
