+++ ai_ref_kits/multimodal_ai_visual_generator/requirements.txt
@@ -1,23 +1,30 @@
---extra-index-url https://download.pytorch.org/whl/cpu
-
-openvino-tokenizers==2024.6.0
+# Core Inference & Optimization
+openvino==2025.1.0
+openvino-genai==2025.1.0
+openvino-tokenizers==2025.1.0
+optimum==1.24.0
 optimum-intel==1.22.0
-numpy==1.26.4
-einops==0.8.0  # For Qwen
-transformers_stream_generator==0.0.5  # For Qwen
-diffusers==0.31.0 # For image generation pipelines
-timm==1.0.11  # For exporting InternVL2
-torchvision==0.20.1  # For visual language models
-transformers==4.48.3 # For Whisper
-nncf==2.15.0
-accelerate==1.3.0
-huggingface_hub==0.28.0
+nncf==2.16.0
+
+# Model + Tokenizer Management
+transformers==4.48.3
+diffusers==0.33.1
+huggingface-hub==0.30.2
+sentencepiece==0.2.0
+protobuf==6.30.2
+peft==0.15.2
+torch==2.7.0
+scipy==1.13.1
+
+# Streamlit Frontend
+streamlit==1.45.0
+pillow==11.2.1
+fpdf==1.7.2
+
+# FastAPI Backend
+fastapi==0.115.12
+uvicorn==0.34.2
+
+# Utility
 opencv-python==4.11.0.86
-pillow==11.1.0
-#openvino GenAI (which also installs openvino packages)
-openvino_genai==2024.6.0
-# for VAD
-sherpa-onnx==1.10.46
-sounddevice==0.5.1
-# for the UI
-PySide6==6.8.1
+numpy==1.26.4
