--- ai_ref_kits/visual_gen_studio/requirements.txt
@@ -1,30 +1,0 @@
-# Core Inference & Optimization
-openvino==2025.1.0
-openvino-genai==2025.1.0
-openvino-tokenizers==2025.1.0
-optimum==1.24.0
-optimum-intel==1.22.0
-nncf==2.16.0
-
-# Model + Tokenizer Management
-transformers==4.48.3
-diffusers==0.33.1
-huggingface-hub==0.30.2
-sentencepiece==0.2.0
-protobuf==6.30.2
-peft==0.15.2
-torch==2.7.0
-scipy==1.13.1
-
-# Streamlit Frontend
-streamlit==1.45.0
-pillow==11.2.1
-fpdf==1.7.2
-
-# FastAPI Backend
-fastapi==0.115.12
-uvicorn==0.34.2
-
-# Utility
-opencv-python==4.11.0.86
-numpy==1.26.4
